requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --extra-index-url=https://packagecloud.io/cytomine-uliege/Cytomine-python-client/pypi/simple
#
--extra-index-url https://packagecloud.io/cytomine-uliege/Cytomine-python-client/pypi/simple

affine==2.3.0
    # via rasterio
aiofiles==0.7.0
    # via pims (setup.py)
aioredis==2.0.0
    # via
    #   fastapi-cache2
    #   pims (setup.py)
amqp==5.0.6
    # via kombu
anyio==3.4.0
    # via starlette
asciitree==0.3.3
    # via zarr
asgiref==3.4.1
    # via uvicorn
async-timeout==4.0.1
    # via aioredis
attrs==21.2.0
    # via rasterio
billiard==3.6.4.0
    # via celery
cachecontrol==0.12.10
    # via cytomine-python-client
celery==5.2.1
    # via pims (setup.py)
certifi==2021.10.8
    # via
    #   rasterio
    #   requests
cffi==1.15.0
    # via pyvips
charset-normalizer==2.0.8
    # via requests
click==8.0.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   cligj
    #   rasterio
    #   uvicorn
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via
    #   celery
    #   rasterio
click-repl==0.2.0
    # via celery
cligj==0.7.2
    # via rasterio
colorama==0.4.4
    # via rich
commonmark==0.9.1
    # via rich
cycler==0.11.0
    # via matplotlib
cytomine-python-client==2.8.3
    # via pims (setup.py)
fastapi==0.70.0
    # via
    #   fastapi-cache2
    #   pims (setup.py)
fastapi-cache2[redis]==0.1.6
    # via pims (setup.py)
fasteners==0.16.3
    # via zarr
fonttools==4.28.2
    # via matplotlib
future==0.18.2
    # via cytomine-python-client
h11==0.12.0
    # via uvicorn
httptools==0.2.0
    # via uvicorn
idna==3.3
    # via
    #   anyio
    #   requests
imagecodecs==2021.11.20
    # via pims (setup.py)
imageio==2.12.0
    # via scikit-image
importlib-metadata==4.8.2
    # via pims (setup.py)
kiwisolver==1.3.2
    # via matplotlib
kombu==5.2.2
    # via celery
matplotlib==3.5.0
    # via scikit-image
msgpack==1.0.3
    # via cachecontrol
networkx==2.6.3
    # via scikit-image
numcodecs==0.9.1
    # via zarr
numpy==1.21.4
    # via
    #   cytomine-python-client
    #   imagecodecs
    #   imageio
    #   matplotlib
    #   numcodecs
    #   opencv-python-headless
    #   pims (setup.py)
    #   pywavelets
    #   rasterio
    #   scikit-image
    #   scipy
    #   snuggs
    #   tifffile
    #   zarr
opencv-python-headless==4.5.4.60
    # via cytomine-python-client
packaging==21.3
    # via
    #   matplotlib
    #   pint
    #   setuptools-scm
pathvalidate==2.5.0
    # via pims (setup.py)
pillow==8.4.0
    # via
    #   cytomine-python-client
    #   imageio
    #   matplotlib
    #   pims (setup.py)
    #   scikit-image
pint==0.18
    # via pims (setup.py)
pkgconfig==1.5.5
    # via pyvips
prompt-toolkit==3.0.22
    # via click-repl
pycparser==2.21
    # via cffi
pydantic==1.8.2
    # via
    #   fastapi
    #   pims (setup.py)
pydicom==2.2.2
    # via pims (setup.py)
pygments==2.10.0
    # via rich
pyparsing==3.0.6
    # via
    #   matplotlib
    #   packaging
    #   snuggs
python-dateutil==2.8.2
    # via
    #   fastapi-cache2
    #   matplotlib
python-dotenv==0.19.2
    # via
    #   pims (setup.py)
    #   uvicorn
python-gdcm==3.0.10
    # via pims (setup.py)
python-multipart==0.0.5
    # via pims (setup.py)
pytz==2021.3
    # via celery
pyvips==2.1.16
    # via pims (setup.py)
pywavelets==1.2.0
    # via scikit-image
pyyaml==6.0
    # via uvicorn
rasterio==1.2.10
    # via pims (setup.py)
requests==2.26.0
    # via
    #   cachecontrol
    #   cytomine-python-client
    #   requests-toolbelt
requests-toolbelt==0.9.1
    # via cytomine-python-client
rich==10.14.0
    # via pims (setup.py)
scikit-image==0.18.3
    # via pims (setup.py)
scipy==1.7.3
    # via scikit-image
setuptools-scm==6.3.2
    # via matplotlib
shapely==1.8.0
    # via
    #   cytomine-python-client
    #   pims (setup.py)
six==1.16.0
    # via
    #   click-repl
    #   cytomine-python-client
    #   fasteners
    #   python-dateutil
    #   python-multipart
sniffio==1.2.0
    # via anyio
snuggs==1.4.7
    # via rasterio
starlette==0.16.0
    # via fastapi
tifffile==2021.6.6
    # via
    #   pims (setup.py)
    #   scikit-image
tomli==1.2.2
    # via setuptools-scm
typing-extensions==4.0.0
    # via
    #   aioredis
    #   async-timeout
    #   pydantic
urllib3==1.26.7
    # via
    #   cytomine-python-client
    #   requests
uvicorn[standard]==0.15.0
    # via
    #   fastapi-cache2
    #   pims (setup.py)
uvloop==0.16.0
    # via uvicorn
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
watchgod==0.7
    # via uvicorn
wcwidth==0.2.5
    # via prompt-toolkit
websockets==10.1
    # via uvicorn
zarr==2.10.3
    # via pims (setup.py)
zipp==3.6.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
